[#7] Add externals builders
Adds support for building externals.

Allows the user to specify a location in the container where external
packages can be found and installed for use with builders. The path
should be a volume mount to a location on the host machine.

In order to be a little more flexible, the file extension for each
platform is now defined in an environment variable in the Dockerfile.