[#63] Add support for building remote repos
Adds 4 new options for building remote repositories:
	- --irods-repo-url
	- --irods-repo-branch
	- --icommands-repo-url
	- --icommands-repo-branch

This allows users to specify a URL and a branch which will be cloned
into the builder rather than using an existing directory created by
a volume mount.

A bonus option was added to only build iCommands: --icommands-only

This requires built iRODS packages to be available at the usual
expected location.